+++ src/btorexp.c
@@ -372,7 +372,7 @@ btor_exp_bv_max_signed (Btor *btor, BtorSortId sort)
 }
 
 BtorNode *
-btor_exp_bv_int (Btor *btor, int32_t i, BtorSortId sort)
+btor_exp_bv_int (Btor *btor, int64_t i, BtorSortId sort)
 {
   assert (btor);
   assert (sort);
@@ -390,7 +390,7 @@ btor_exp_bv_int (Btor *btor, int32_t i, BtorSortId sort)
 }
 
 BtorNode *
-btor_exp_bv_unsigned (Btor *btor, uint32_t u, BtorSortId sort)
+btor_exp_bv_unsigned (Btor *btor, uint64_t u, BtorSortId sort)
 {
   assert (btor);
   assert (sort);
